+++ news/goodbye-python-3.5.rst
@@ -1,0 +1,26 @@
+**Added:**
+
+* <news item>
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* Compatibility with Python 3.5 has been removed as well as all related code. In
+  particular xonsh.inspector does not defined ``getouterframes`` anymore, use
+  ``inspect.getouterframe`` directly.
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
+
