+++ xonsh-in-docker.py
@@ -9,7 +9,7 @@ program_description = """Build and run Xonsh in a fresh, controlled
 parser = argparse.ArgumentParser(description=program_description)
 
 parser.add_argument("env", nargs="*", default=[], metavar="ENV=value")
-parser.add_argument("--python", "-p", default="3.5", metavar="python_version")
+parser.add_argument("--python", "-p", default="3.6", metavar="python_version")
 parser.add_argument("--pypy", default=None, metavar="pypy_version")
 parser.add_argument("--ptk", "-t", default="2.0.4", metavar="ptk_version")
 parser.add_argument("--keep", action="store_true")
