+++ xonsh/lexer.py
@@ -106,7 +106,7 @@ def token_map():
     tm[NEWLINE] = "NEWLINE"
     tm[INDENT] = "INDENT"
     tm[DEDENT] = "DEDENT"
-    if (3, 5, 0) <= PYTHON_VERSION_INFO < (3, 7, 0):
+    if PYTHON_VERSION_INFO < (3, 7, 0):
         from xonsh.tokenize import ASYNC, AWAIT
 
         tm[ASYNC] = "ASYNC"
