+++ xonsh/parser.py
@@ -8,8 +8,6 @@ from xonsh.platform import PYTHON_VERSION_INFO
 def Parser():
     if PYTHON_VERSION_INFO > (3, 8):
         from xonsh.parsers.v38 import Parser as p
-    elif PYTHON_VERSION_INFO > (3, 6):
+    else:
         from xonsh.parsers.v36 import Parser as p
-    elif PYTHON_VERSION_INFO > (3, 5):
-        from xonsh.parsers.v35 import Parser as p
     return p
